pco_api v0.2.0 PcoApi.Query

A set of functions to assist in composing queries to the API.

Functions

after_record(value)

Creates a List of after params to pass on to the api call.

Returns a List of Tuples.

Examples:

iex> PcoApi.Query.per_page(4) |> PcoApi.Query.after_record(4)
[{"after", "4"}, {"per_page", "4"}]

after_record(params, value)

per_page(value)

Creates a list of per_page params to pass on to the api call.

Returns a List of Tuples.

Examples:

iex> PcoApi.Query.where(first_name: "John") |> PcoApi.Query.per_page(3)
[{"per_page", "3"}, {"where[first_name]", "John"}]

per_page(params, value)

where(param)

Creates a list of where params to pass onto the api call. Returns a List of Tuples.

Examples

iex> PcoApi.Query.where(first_name: "Geoffrey")
[{"where[first_name]", "Geoffrey"}]

iex> PcoApi.Query.where(first_name: "Geoffrey") |> PcoApi.Query.where(last_name: "Lessel")
[{"where[last_name]", "Lessel"}, {"where[first_name]", "Geoffrey"}]

where(params, param)